service-c/requirements.txt
fastapi
uvicorn
pydantic
requests
dotenv
mysql-connector-python